+++ server/cohort/models.py
@@ -1,5 +1,4 @@
 from django.db import models
-from members.models import User
 
 
 class Cohort(models.Model):
@@ -24,15 +23,3 @@ class Cohort(models.Model):
 
     def __str__(self):
         return f"{self.name} ({self.get_level_display()})"
-
-
-class CohortStats(models.Model):
-    member = models.ForeignKey(
-        User, on_delete=models.CASCADE, related_name="cohort_member_stats"
-    )
-    cohort = models.ForeignKey(Cohort, on_delete=models.CASCADE)
-    applications = models.IntegerField(default=0)
-    onlineAssessments = models.IntegerField(default=0)
-    interviews = models.IntegerField(default=0)
-    offers = models.IntegerField(default=0)
-    dailyChecks = models.IntegerField(default=0)
