+++ server/cohort/serializers.py
@@ -1,7 +1,7 @@
 from rest_framework import serializers
 
 from members.serializers import UserSerializer
-from .models import Cohort, CohortStats
+from .models import Cohort
 
 
 class CohortSerializer(serializers.ModelSerializer):
@@ -16,20 +16,3 @@ class CohortHydratedSerializer(serializers.ModelSerializer):
     class Meta:
         model = Cohort
         fields = ["id", "name", "members"]
-
-
-class CohortStatsSerializer(serializers.ModelField):
-    member = UserSerializer(read_only=True)
-    cohort = CohortSerializer
-
-    class Meta:
-        model = CohortStats
-        fields = [
-            "member",
-            "cohort",
-            "applications",
-            "onlineAssessments",
-            "interviews",
-            "offers",
-            "dailyChecks",
-        ]
